Set user icon to default when updating profile without photo
If the user photo was deleted in the Myself contact, the change
was not reflected neither in Settings/Users nor in expanded
status bar. This was because the ProfileUpdateReceiver did not
handle the case where the updated image was null (deleted).
Fixed by assigning default user icon in UserManager if no photo
found in the profile update.

Bug: 28031914

Change-Id: I8a3cad52ce126fc9eb3a10fa7565814922bd4e4c
Signed-off-by: placiano <placiano80@gmail.com>