Settings: reflect/inform user of disabled wakelock appop
If an app is marked as "allow-in-power-save" (loaded in as part of
SystemConfig), then this app is not susceptible to battery saving
optimizations and is required to be active for the system to function
properly.

This behavior is enforced on the PowerManagerService side when
requesting wake locks - inform the user visually as well.

Ticket: CYNGNOS-1258

Change-Id: I1c99de703012a09ddbde19838171b7e5659f2523
Signed-off-by: Roman Birg <roman@cyngn.com>
Signed-off-by: placiano <placiano80@gmail.com>