Prevent crash if weather settings activity can't be found
Each weather provider service may define a settings activity which
will be automatically launched when the user selects the provider.
If such activity can't be found, an ActivityNotFoundException is
thrown. This patch adds a try/catch block to avoid crashing the
settings app if the activity can't be found

A toast will be displayed to notify the user. Otherwise, the user
would never know why the settings gear is doing nothing

Change-Id: I3c3d54ce8f9131353077c34a2f821ec6b0fc0268
TICKET: CYNGNOS-2478
Signed-off-by: placiano <placiano80@gmail.com>